LPS-57393 Move MVC actions into module
