--- portal-impl/src/META-INF/announcements-spring.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-
-<beans
-	default-destroy-method="destroy"
-	default-init-method="afterPropertiesSet"
-	xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
->
-	<bean id="com.liferay.portlet.announcements.action.EditEntryMVCActionCommand" class="com.liferay.portlet.announcements.action.EditEntryMVCActionCommand" />
-	<bean id="com.liferay.portlet.announcements.action.EditEntryMVCRenderCommand" class="com.liferay.portlet.announcements.action.EditEntryMVCRenderCommand" />
-	<bean id="com.liferay.portlet.announcements.action.PreviewEntryMVCActionCommand" class="com.liferay.portlet.announcements.action.PreviewEntryMVCActionCommand" />
-	<bean id="com.liferay.portlet.announcements.action.PreviewEntryMVCRenderCommand" class="com.liferay.portlet.announcements.action.PreviewEntryMVCRenderCommand" />
-	<bean id="com.liferay.portlet.announcements.action.ViewMVCRenderCommand" class="com.liferay.portlet.announcements.action.ViewMVCRenderCommand" />
-</beans>
