+++ portal-impl/src/portal.properties
@@ -830,7 +830,6 @@
         META-INF/messaging-core-spring.xml,\
         META-INF/messaging-misc-spring.xml,\
         \
-        META-INF/announcements-spring.xml,\
         META-INF/audit-spring.xml,\
         META-INF/cluster-spring.xml,\
         META-INF/comment-spring.xml,\
